[REFACTOR]#160 : Message 상태 및 ViewModel 중 Storage 관련 코드 StorageViewModel / StorageIntent 로 변경
