+++ feature/message/src/main/kotlin/com/bff/wespot/message/state/MessageSideEffect.kt
@@ -1,8 +1,3 @@
 package com.bff.wespot.message.state
 
-import com.bff.wespot.model.ToastState
-
-sealed class MessageSideEffect {
-    data class ShowToast(val toastState: ToastState) : MessageSideEffect()
-    data object ShowMessageDialog : MessageSideEffect()
-}
+sealed class MessageSideEffect
