+++ feature/message/src/main/kotlin/com/bff/wespot/message/state/MessageUiState.kt
@@ -1,30 +1,15 @@
 package com.bff.wespot.message.state
 
-import androidx.paging.PagingData
-import com.bff.wespot.message.model.ClickedMessageUiModel
-import com.bff.wespot.message.model.MessageOptionType
+import com.bff.wespot.message.common.DEFAULT_MESSAGE_RECEIVE_TIME
+import com.bff.wespot.message.common.DEFAULT_MESSAGE_START_TIME
 import com.bff.wespot.message.model.TimePeriod
-import com.bff.wespot.model.message.request.MessageType
 import com.bff.wespot.model.message.response.Message
 import com.bff.wespot.model.message.response.MessageStatus
-import com.bff.wespot.model.message.response.ReceivedMessage
-import com.bff.wespot.model.message.response.SentMessage
-import com.bff.wespot.model.user.response.Profile
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 
 data class MessageUiState(
     val timePeriod: TimePeriod = TimePeriod.DAWN_TO_EVENING,
     val messageStatus: MessageStatus = MessageStatus(false, -1, -1),
-    val myProfile: Profile = Profile(),
-    val receivedMessageList: Flow<PagingData<ReceivedMessage>> = flow { },
-    val sentMessageList: Flow<PagingData<SentMessage>> = flow { },
-    val clickedMessage: ClickedMessageUiModel = ClickedMessageUiModel(),
-    val optionButtonClickedMessageId: Int = -1,
-    val optionButtonClickedMessageType: MessageType = MessageType.RECEIVED,
-    val messageOptionType: MessageOptionType = MessageOptionType.DELETE,
+    val messageStartTime: String = DEFAULT_MESSAGE_START_TIME,
+    val messageReceiveTime: String = DEFAULT_MESSAGE_RECEIVE_TIME,
     val reservedMessageList: List<Message> = listOf(),
-    val isLoading: Boolean = false,
-    val messageStartTime: String = "17:00",
-    val messageReceiveTime: String = "22:00",
 )
