+++ feature/message/src/main/kotlin/com/bff/wespot/message/viewmodel/MessageViewModel.kt
@@ -1,29 +1,16 @@
 package com.bff.wespot.message.viewmodel
 
 import androidx.lifecycle.viewModelScope
-import androidx.paging.cachedIn
 import com.bff.wespot.base.BaseViewModel
-import com.bff.wespot.designsystem.component.indicator.WSToastType
-import com.bff.wespot.domain.repository.BasePagingRepository
-import com.bff.wespot.domain.repository.CommonRepository
 import com.bff.wespot.domain.repository.RemoteConfigRepository
 import com.bff.wespot.domain.repository.message.MessageRepository
 import com.bff.wespot.domain.repository.message.MessageStorageRepository
 import com.bff.wespot.domain.util.RemoteConfigKey
-import com.bff.wespot.message.R
-import com.bff.wespot.message.model.ClickedMessageUiModel
-import com.bff.wespot.message.model.MessageOptionType
 import com.bff.wespot.message.model.TimePeriod
 import com.bff.wespot.message.model.getCurrentTimePeriod
 import com.bff.wespot.message.state.MessageAction
 import com.bff.wespot.message.state.MessageSideEffect
 import com.bff.wespot.message.state.MessageUiState
-import com.bff.wespot.model.ToastState
-import com.bff.wespot.model.common.Paging
-import com.bff.wespot.model.common.ReportType
-import com.bff.wespot.model.message.request.MessageType
-import com.bff.wespot.model.message.response.ReceivedMessage
-import com.bff.wespot.model.message.response.SentMessage
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineStart
 import kotlinx.coroutines.Job
@@ -35,19 +22,14 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.orbitmvi.orbit.ContainerHost
 import org.orbitmvi.orbit.syntax.simple.intent
-import org.orbitmvi.orbit.syntax.simple.postSideEffect
 import org.orbitmvi.orbit.syntax.simple.reduce
 import org.orbitmvi.orbit.viewmodel.container
-import timber.log.Timber
 import javax.inject.Inject
 
 @HiltViewModel
 class MessageViewModel @Inject constructor(
     private val messageRepository: MessageRepository,
     private val messageStorageRepository: MessageStorageRepository,
-    private val messageReceivedRepository: BasePagingRepository<ReceivedMessage, Paging<ReceivedMessage>>,
-    private val messageSentRepository: BasePagingRepository<SentMessage, Paging<SentMessage>>,
-    private val commonRepository: CommonRepository,
     remoteConfigRepository: RemoteConfigRepository,
 ) : BaseViewModel(), ContainerHost<MessageUiState, MessageSideEffect> {
     override val container = container<MessageUiState, MessageSideEffect>(
@@ -99,36 +81,6 @@ class MessageViewModel @Inject constructor(
     fun onAction(action: MessageAction) {
         when (action) {
             is MessageAction.OnHomeScreenEntered -> handleHomeScreenEntered()
-            is MessageAction.OnStorageChipSelected -> {
-                when (action.messageType) {
-                    MessageType.SENT -> {
-                        getMessageStatus()
-                        getSentMessageList()
-                    }
-                    MessageType.RECEIVED -> getReceivedMessageList()
-                }
-            }
-            is MessageAction.OnMessageStorageScreenOpened -> {
-                handleMessageStorageScreenOpened(action.messageId, action.type)
-            }
-            is MessageAction.OnSentMessageClicked ->
-                handleSentMessageClicked(action.message)
-            is MessageAction.OnReceivedMessageClicked ->
-                handleReceivedMessageClicked(action.message)
-            is MessageAction.OnOptionButtonClicked ->
-                handleOptionButtonClicked(action.messageId, action.messageType)
-            is MessageAction.OnOptionBottomSheetClicked -> {
-                handleOptionBottomSheetClicked(action.messageOptionType)
-            }
-            MessageAction.OnMessageDeleteButtonClicked -> {
-                handleDeleteMessageButtonClicked()
-            }
-            MessageAction.OnMessageReportButtonClicked -> {
-                handleReportMessageButtonClicked()
-            }
-            MessageAction.OnMessageBlockButtonClicked -> {
-                handleBlockMessageButtonClicked()
-            }
             MessageAction.OnReservedMessageScreenEntered -> handleReservedMessageScreenEntered()
         }
     }
@@ -184,166 +136,6 @@ class MessageViewModel @Inject constructor(
                             messageStatus = messageStatus,
                         )
                     }
-                }
-        }
-    }
-
-    private fun getReceivedMessageList() = intent {
-        viewModelScope.launch(coroutineDispatcher) {
-            runCatching {
-                reduce {
-                    state.copy(
-                        receivedMessageList = messageReceivedRepository.fetchResultStream()
-                            .cachedIn(viewModelScope),
-                    )
-                }
-            }.onFailure { exception ->
-                Timber.e(exception)
-            }
-        }
-    }
-
-    private fun getSentMessageList() = intent {
-        viewModelScope.launch(coroutineDispatcher) {
-            runCatching {
-                reduce {
-                    state.copy(
-                        sentMessageList = messageSentRepository.fetchResultStream()
-                            .cachedIn(viewModelScope),
-                    )
-                }
-            }.onFailure { exception ->
-                Timber.e(exception)
-            }
-        }
-    }
-
-    private fun handleMessageStorageScreenOpened(messageId: Int, type: MessageType) = intent {
-        reduce { state.copy(isLoading = true) }
-        viewModelScope.launch {
-            messageRepository.getMessage(messageId)
-                .onSuccess { message ->
-                    when (type) {
-                        MessageType.SENT -> handleSentMessageClicked(message.toSentMessage())
-                        MessageType.RECEIVED -> handleReceivedMessageClicked(
-                            message.toReceivedMessage(),
-                        )
-                    }
-                    reduce { state.copy(isLoading = false) }
-                    postSideEffect(MessageSideEffect.ShowMessageDialog)
-                }
-                .onFailure {
-                    reduce { state.copy(isLoading = false) }
-                }
-        }
-    }
-
-    private fun handleReceivedMessageClicked(message: ReceivedMessage) = intent {
-        reduce {
-            state.copy(
-                clickedMessage = ClickedMessageUiModel(
-                    content = message.content,
-                    sender = message.senderName,
-                    receiver = message.receiver.toDescription(),
-                ),
-            )
-        }
-
-        if (message.isRead.not()) {
-            updateMessageReadStatus(messageId = message.id)
-        }
-    }
-
-    private fun handleSentMessageClicked(message: SentMessage) = intent {
-        reduce {
-            state.copy(
-                clickedMessage = ClickedMessageUiModel(
-                    content = message.content,
-                    sender = message.senderName,
-                    receiver = message.receiver.toDescription(),
-                ),
-            )
-        }
-    }
-
-    private fun handleOptionButtonClicked(messageId: Int, messageType: MessageType) = intent {
-        reduce {
-            state.copy(
-                optionButtonClickedMessageId = messageId,
-                optionButtonClickedMessageType = messageType,
-            )
-        }
-    }
-
-    private fun handleOptionBottomSheetClicked(messageOptionType: MessageOptionType) = intent {
-        reduce {
-            state.copy(messageOptionType = messageOptionType)
-        }
-    }
-
-    private fun updateMessageReadStatus(messageId: Int) {
-        viewModelScope.launch {
-            messageStorageRepository.updateMessageReadStatus(messageId)
-                .onSuccess {
-                    getReceivedMessageList()
-                }
-        }
-    }
-
-    private fun handleDeleteMessageButtonClicked() = intent {
-        viewModelScope.launch {
-            messageStorageRepository.deleteMessage(state.optionButtonClickedMessageId)
-                .onSuccess {
-                    when (state.optionButtonClickedMessageType) {
-                        MessageType.RECEIVED -> getReceivedMessageList()
-                        MessageType.SENT -> getSentMessageList()
-                    }
-
-                    postSideEffect(
-                        MessageSideEffect.ShowToast(
-                            ToastState(
-                                show = true,
-                                message = R.string.delete_done,
-                                type = WSToastType.Success,
-                            ),
-                        ),
-                    )
-                }
-        }
-    }
-
-    private fun handleReportMessageButtonClicked() = intent {
-        viewModelScope.launch {
-            commonRepository.sendReport(ReportType.MESSAGE, state.optionButtonClickedMessageId)
-                .onSuccess {
-                    postSideEffect(
-                        MessageSideEffect.ShowToast(
-                            ToastState(
-                                show = true,
-                                message = R.string.report_done,
-                                type = WSToastType.Success,
-                            ),
-                        ),
-                    )
-                    getReceivedMessageList()
-                }
-        }
-    }
-
-    private fun handleBlockMessageButtonClicked() = intent {
-        viewModelScope.launch {
-            messageStorageRepository.blockMessage(state.optionButtonClickedMessageId)
-                .onSuccess {
-                    postSideEffect(
-                        MessageSideEffect.ShowToast(
-                            ToastState(
-                                show = true,
-                                message = R.string.block_done,
-                                type = WSToastType.Success,
-                            ),
-                        ),
-                    )
-                    getReceivedMessageList()
                 }
         }
     }
